requirements.txt
streamlit==1.38.0
python-docx==1.1.2
PyMuPDF==1.24.10
